# order-service/internal/handler/handler.go
package handler

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/mnntn/ecommerce-project/order-service/internal/service"
)

type OrderHandler struct {
	service *service.OrderService
}

func NewOrderHandler(service *service.OrderService) *OrderHandler {
	return &OrderHandler{
		service: service,
	}
}

func (h *OrderHandler) Start(addr string) error {
	r := gin.Default()

	r.POST("/orders", h.createOrder)
	r.GET("/orders/:id", h.getOrder)

	return r.Run(addr)
}

func (h *OrderHandler) createOrder(c *gin.Context) {
	var req struct {
		UserID string `json:"user_id" binding:"required"`
	}

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	order, err := h.service.CreateOrder(c.Request.Context(), req.UserID)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusCreated, order)
}

func (h *OrderHandler) getOrder(c *gin.Context) {
	id := c.Param("id")

	order, err := h.service.GetOrder(c.Request.Context(), id)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, order)
}