# order-service/internal/service/requests.go
package service

type CreateOrderRequest struct {
	UserID string            `json:"user_id"`
	Items  []CreateOrderItem `json:"items"`
}

type CreateOrderItem struct {
	ProductID int64 `json:"product_id"`
	Quantity  int   `json:"quantity"`
}